requirements.txt
Django>=3.2.5,<3.3
psycopg2==2.9.5
psycopg2-binary==2.9.6
import-export==0.2.67.dev6
django-import-export==2.7.1
xlrd2>=1.3.4,<1.4
openpyxl>=3.0.9,<3.1
Pillow==10.3.0
djangorestframework>=3.15.1,<3.16
boto3==1.34.127
django-storages==1.14.3
uWSGI==2.0.26
django-cors-headers>=3.13.0,<3.14
graphene-django>=2.15.0,<2.16
graphene-file-upload>=1.3.0,<1.4
django-graphql-jwt>=0.3.4,<0.4
PyJWT==2.0.0
#dj-database-url
# supabase==2.13.0
#django-tailwind
#sweetify